Add string to float conversion
